move Dynamic Rectors to its own namespaces
