+++ tests/Rector/Class_/ClassReplacerRector/Correct/correct.php.inc
