+++ tests/Rector/Class_/ClassReplacerRector/Correct/correct2.php.inc
