+++ tests/Rector/Class_/ClassReplacerRector/Correct/correct3.php.inc
