+++ tests/Rector/Class_/ClassReplacerRector/Wrong/wrong.php.inc
