+++ tests/Rector/Class_/ClassReplacerRector/Wrong/wrong2.php.inc
