+++ tests/Rector/Class_/ClassReplacerRector/Wrong/wrong3.php.inc
