+++ tests/Rector/Property/PropertyNameReplacerRector/Correct/correct.php.inc
