+++ tests/Rector/Property/PropertyNameReplacerRector/Wrong/wrong.php.inc
