+++ tests/Rector/Property/PropertyToMethodRector/Correct/correct.php.inc
