+++ tests/Rector/Property/PropertyToMethodRector/Wrong/wrong.php.inc
